Split server state into feature modules
