+++ crates/presenter-server/src/state/presentations.rs
@@ -1,0 +1,239 @@
+use std::{collections::HashMap, sync::Arc};
+
+use anyhow::{anyhow, Result};
+use presenter_core::{
+    LibraryId, Presentation, PresentationId, Slide, SlideContent, SlideGroup, SlideId, SlideText,
+};
+
+use super::{blank_slide_content, AppState};
+
+impl AppState {
+    pub(super) async fn presentation_from_cache(
+        &self,
+        presentation_id: PresentationId,
+    ) -> Result<Arc<Presentation>> {
+        if let Some(cached) = {
+            let guard = self.presentation_cache.read().await;
+            guard.get(&presentation_id).cloned()
+        } {
+            return Ok(cached);
+        }
+        let detail = self
+            .repository
+            .fetch_presentation_detail(presentation_id)
+            .await?;
+        let Some((_, _, presentation)) = detail else {
+            return Err(anyhow!("presentation not found"));
+        };
+        let arc = Arc::new(presentation);
+        let mut guard = self.presentation_cache.write().await;
+        guard.insert(presentation_id, arc.clone());
+        Ok(arc)
+    }
+
+    pub(super) async fn cache_presentation_ref(&self, presentation: &Presentation) {
+        let mut guard = self.presentation_cache.write().await;
+        guard.insert(presentation.id, Arc::new(presentation.clone()));
+    }
+
+    pub(super) async fn cache_presentation_value(&self, presentation: Presentation) {
+        let mut guard = self.presentation_cache.write().await;
+        guard.insert(presentation.id, Arc::new(presentation));
+    }
+
+    pub async fn presentation_detail(
+        &self,
+        presentation_id: PresentationId,
+    ) -> Result<Option<(LibraryId, String, Presentation)>> {
+        let detail = self
+            .repository
+            .fetch_presentation_detail(presentation_id)
+            .await?;
+        if let Some((library_id, library_name, presentation)) = detail {
+            self.cache_presentation_ref(&presentation).await;
+            Ok(Some((library_id, library_name, presentation)))
+        } else {
+            Ok(None)
+        }
+    }
+
+    pub async fn update_slide_content(
+        &self,
+        presentation_id: PresentationId,
+        slide_id: SlideId,
+        main: String,
+        translation: String,
+        stage: String,
+        group: Option<String>,
+    ) -> Result<Slide> {
+        let presentation_arc = self.presentation_from_cache(presentation_id).await?;
+        let presentation = presentation_arc.as_ref();
+
+        let existing_slide = presentation
+            .slides
+            .iter()
+            .find(|slide| slide.id == slide_id)
+            .ok_or_else(|| anyhow!("slide not found"))?
+            .clone();
+
+        let main_text = SlideText::new(main).map_err(|err| anyhow!(err))?;
+        let translation_text = SlideText::new(translation).map_err(|err| anyhow!(err))?;
+        let stage_text = SlideText::new(stage).map_err(|err| anyhow!(err))?;
+        let group = group.and_then(|value| {
+            let trimmed = value.trim();
+            if trimmed.is_empty() {
+                None
+            } else {
+                Some(SlideGroup::new(trimmed.to_string()))
+            }
+        });
+
+        let content = SlideContent::new(
+            main_text.clone(),
+            translation_text.clone(),
+            stage_text.clone(),
+            group.clone(),
+        );
+        let updated_slide = Slide::new(existing_slide.order, content.clone()).with_id(slide_id);
+
+        self.repository
+            .update_slide_content(presentation_id, slide_id, &content)
+            .await?;
+
+        let mut updated_presentation = presentation.clone();
+        if let Some(slot) = updated_presentation
+            .slides
+            .iter_mut()
+            .find(|slide| slide.id == slide_id)
+        {
+            *slot = updated_slide.clone();
+        }
+        self.cache_presentation_value(updated_presentation).await;
+
+        self.broadcast_stage_snapshots().await?;
+
+        Ok(updated_slide)
+    }
+
+    pub async fn insert_blank_slide(
+        &self,
+        presentation_id: PresentationId,
+        position: Option<u32>,
+    ) -> Result<Vec<Slide>> {
+        let presentation_arc = self.presentation_from_cache(presentation_id).await?;
+        let presentation = presentation_arc.as_ref();
+        let mut slides = presentation.slides.clone();
+        let insert_at = position
+            .map(|value| value as usize)
+            .unwrap_or(slides.len())
+            .min(slides.len());
+        slides.insert(insert_at, Slide::new(0, blank_slide_content()));
+        Self::reindex_slides(&mut slides);
+        self.repository
+            .replace_presentation_slides(presentation_id, &slides)
+            .await?;
+        self.reconcile_stage_state_after_edit(presentation_id, &slides)
+            .await?;
+        let mut updated_presentation = presentation.clone();
+        updated_presentation.slides = slides.clone();
+        self.cache_presentation_value(updated_presentation).await;
+        self.broadcast_stage_snapshots().await?;
+        Ok(slides)
+    }
+
+    pub async fn duplicate_slide(
+        &self,
+        presentation_id: PresentationId,
+        slide_id: SlideId,
+    ) -> Result<Vec<Slide>> {
+        let presentation_arc = self.presentation_from_cache(presentation_id).await?;
+        let presentation = presentation_arc.as_ref();
+        let mut slides = presentation.slides.clone();
+        let index = slides
+            .iter()
+            .position(|slide| slide.id == slide_id)
+            .ok_or_else(|| anyhow!("slide not found"))?;
+        let source = slides[index].clone();
+        slides.insert(index + 1, Slide::new(0, source.content.clone()));
+        Self::reindex_slides(&mut slides);
+        self.repository
+            .replace_presentation_slides(presentation_id, &slides)
+            .await?;
+        self.reconcile_stage_state_after_edit(presentation_id, &slides)
+            .await?;
+        let mut updated_presentation = presentation.clone();
+        updated_presentation.slides = slides.clone();
+        self.cache_presentation_value(updated_presentation).await;
+        self.broadcast_stage_snapshots().await?;
+        Ok(slides)
+    }
+
+    pub async fn delete_slide(
+        &self,
+        presentation_id: PresentationId,
+        slide_id: SlideId,
+    ) -> Result<Vec<Slide>> {
+        let presentation_arc = self.presentation_from_cache(presentation_id).await?;
+        let presentation = presentation_arc.as_ref();
+        let mut slides = presentation.slides.clone();
+        let index = slides
+            .iter()
+            .position(|slide| slide.id == slide_id)
+            .ok_or_else(|| anyhow!("slide not found"))?;
+        slides.remove(index);
+        if slides.is_empty() {
+            slides.push(Slide::new(0, blank_slide_content()));
+        }
+        Self::reindex_slides(&mut slides);
+        self.repository
+            .replace_presentation_slides(presentation_id, &slides)
+            .await?;
+        self.reconcile_stage_state_after_edit(presentation_id, &slides)
+            .await?;
+        let mut updated_presentation = presentation.clone();
+        updated_presentation.slides = slides.clone();
+        self.cache_presentation_value(updated_presentation).await;
+        self.broadcast_stage_snapshots().await?;
+        Ok(slides)
+    }
+
+    pub async fn reorder_slides(
+        &self,
+        presentation_id: PresentationId,
+        order: Vec<SlideId>,
+    ) -> Result<Vec<Slide>> {
+        let presentation_arc = self.presentation_from_cache(presentation_id).await?;
+        let presentation = presentation_arc.as_ref();
+        let mut map = HashMap::new();
+        for slide in presentation.slides.clone() {
+            map.insert(slide.id, slide);
+        }
+        if order.len() != map.len() {
+            return Err(anyhow!("slide order length mismatch"));
+        }
+        let mut slides = Vec::with_capacity(order.len());
+        for id in order {
+            let slide = map
+                .remove(&id)
+                .ok_or_else(|| anyhow!("unknown slide in reorder request"))?;
+            slides.push(slide);
+        }
+        Self::reindex_slides(&mut slides);
+        self.repository
+            .replace_presentation_slides(presentation_id, &slides)
+            .await?;
+        self.reconcile_stage_state_after_edit(presentation_id, &slides)
+            .await?;
+        let mut updated_presentation = presentation.clone();
+        updated_presentation.slides = slides.clone();
+        self.cache_presentation_value(updated_presentation).await;
+        self.broadcast_stage_snapshots().await?;
+        Ok(slides)
+    }
+
+    pub(super) fn reindex_slides(slides: &mut Vec<Slide>) {
+        for (index, slide) in slides.iter_mut().enumerate() {
+            slide.order = index as u32;
+        }
+    }
+}
