+++ crates/presenter-server/src/state/tests.rs
@@ -1,6 +1,10 @@
 use super::*;
 use crate::live::LiveEvent;
-use presenter_core::{bible::BibleIngestionBatch, BiblePassage, BibleTranslation, TimerState};
+use chrono::Utc;
+use presenter_core::{
+    bible::BibleIngestionBatch, BiblePassage, BibleReference, BibleTranslation, Library, LibraryId,
+    Presentation, Slide, SlideContent, SlideId, SlideText, TimerCommand, TimerState,
+};
 
 #[tokio::test]
 async fn seeded_state_contains_library() {
